auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
asyncpg==0.29.0
pydantic-settings==2.3.4
pydantic[email]==2.8.2
python-dotenv==1.0.1
boto3==1.34.144
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
pyotp==2.9.0
httpx==0.27.0

# dev / test
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
